added: subgraph explorer: uniswap, compound, synthetix, balancer etc
