+++ packages/react-app/src/views/index.js
@@ -1,3 +1,4 @@
 export { default as ExampleUI } from "./ExampleUI";
 export { default as Hints } from "./Hints";
 export { default as Subgraph } from "./Subgraph";
+export { default as DexSubgraphExplorer } from "./DexSubgraphExplorer";
